refactor(ivy): refactor more files in DI to prepare it for bazel packages (#28098)
PR Close #28098
