+++ packages/compiler/test/BUILD.bazel
@@ -41,6 +41,7 @@ ts_library(
         "//packages/compiler/test/ml_parser/util",
         "//packages/compiler/testing",
         "//packages/core",
+        "//packages/core/src/compiler",
         "//packages/core/testing",
         "//packages/platform-browser",
         "//packages/platform-browser-dynamic",
