+++ packages/core/src/compiler/BUILD.bazel
@@ -1,0 +1,20 @@
+package(default_visibility = [
+    "//packages/compiler/test:__pkg__",
+    "//packages/core:__subpackages__",
+    "//tools/public_api_guard:__pkg__",
+])
+
+load("//tools:defaults.bzl", "ts_library")
+
+ts_library(
+    name = "compiler",
+    srcs = glob(
+        [
+            "**/*.ts",
+        ],
+    ),
+    module_name = "@angular/core/compiler",
+    deps = [
+        "//packages/core/src/util",
+    ],
+)
