+++ packages/core/src/compiler/compiler_facade.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {global} from '../../util/global';
+import {global} from '../util/global';
 import {CompilerFacade, ExportedCompilerFacade} from './compiler_facade_interface';
 export * from './compiler_facade_interface';
