+++ packages/core/src/di/injectable.ts
@@ -7,11 +7,10 @@
  */
 
 import {Type} from '../interface/type';
-import {compileInjectable as render3CompileInjectable} from '../render3/jit/injectable';
 import {TypeDecorator, makeDecorator} from '../util/decorators';
-
-import {ClassSansProvider, ConstructorSansProvider, ExistingSansProvider, FactorySansProvider, StaticClassSansProvider, ValueSansProvider,} from './interface/provider';
-import {InjectableType, defineInjectable, getInjectableDef, InjectableDef} from './interface/defs';
+import {InjectableDef, InjectableType, defineInjectable, getInjectableDef} from './interface/defs';
+import {ClassSansProvider, ConstructorSansProvider, ExistingSansProvider, FactorySansProvider, StaticClassSansProvider, ValueSansProvider} from './interface/provider';
+import {compileInjectable as render3CompileInjectable} from './jit/injectable';
 import {convertInjectableProviderToFactory} from './util';
