+++ packages/core/src/error_handler.ts
@@ -7,7 +7,6 @@
  */
 
 import {getDebugContext, getErrorLogger, getOriginalError} from './errors';
-import {ERROR_ORIGINAL_ERROR} from './util/errors';
 
 
 
@@ -77,12 +76,4 @@ export class ErrorHandler {
 
     return e;
   }
-}
-
-export function wrappedError(message: string, originalError: any): Error {
-  const msg =
-      `${message} caused by: ${originalError instanceof Error ? originalError.message: originalError }`;
-  const error = Error(msg);
-  (error as any)[ERROR_ORIGINAL_ERROR] = originalError;
-  return error;
 }
