+++ packages/core/src/metadata/di.ts
@@ -8,7 +8,7 @@
 
 import {InjectionToken} from '../di/injection_token';
 import {Type} from '../interface/type';
-import {makeParamDecorator, makePropDecorator} from '../util/decorators';
+import {makePropDecorator} from '../util/decorators';
 
 /**
  * This token can be used to create a virtual provider that will populate the
@@ -47,71 +47,6 @@ import {makeParamDecorator, makePropDecorator} from '../util/decorators';
  */
 export const ANALYZE_FOR_ENTRY_COMPONENTS = new InjectionToken<any>('AnalyzeForEntryComponents');
 
-/**
- * Type of the Attribute decorator / constructor function.
- *
- * @publicApi
- */
-export interface AttributeDecorator {
-  /**
-   * Specifies that a constant attribute value should be injected.
-   *
-   * The directive can inject constant string literals of host element attributes.
-   *
-   * @usageNotes
-   * ### Example
-   *
-   * Suppose we have an `<input>` element and want to know its `type`.
-   *
-   * ```html
-   * <input type="text">
-   * ```
-   *
-   * A decorator can inject string literal `text` like so:
-   *
-   * {@example core/ts/metadata/metadata.ts region='attributeMetadata'}
-   *
-   * ### Example as TypeScript Decorator
-   *
-   * {@example core/ts/metadata/metadata.ts region='attributeFactory'}
-   *
-   * ### Example as ES5 annotation
-   *
-   * ```
-   * var MyComponent = function(title) {
-   *   ...
-   * };
-   *
-   * MyComponent.annotations = [
-   *   new ng.Component({...})
-   * ]
-   * MyComponent.parameters = [
-   *   [new ng.Attribute('title')]
-   * ]
-   * ```
-   *
-   * @publicApi
-   */
-  (name: string): any;
-  new (name: string): Attribute;
-}
-
-
-/**
- * Type of the Attribute metadata.
- *
- * @publicApi
- */
-export interface Attribute { attributeName?: string; }
-
-/**
- * Attribute decorator and metadata.
- *
- * @Annotation
- * @publicApi
- */
-export const Attribute: AttributeDecorator =
-    makeParamDecorator('Attribute', (attributeName?: string) => ({attributeName}));
 
 /**
  * Type of the Query metadata.
