+++ packages/core/src/reflection/BUILD.bazel
@@ -1,0 +1,20 @@
+package(default_visibility = [
+    "//packages/core:__subpackages__",
+    "//tools/public_api_guard:__pkg__",
+])
+
+load("//tools:defaults.bzl", "ts_library")
+
+ts_library(
+    name = "reflection",
+    srcs = glob(
+        [
+            "**/*.ts",
+        ],
+    ),
+    module_name = "@angular/core/reflection",
+    deps = [
+        "//packages/core/src/interface",
+        "//packages/core/src/util",
+    ],
+)
