+++ packages/core/src/render3/onchanges_util.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {SimpleChange, SimpleChanges} from '../change_detection/simple_change';
+import {SimpleChange, SimpleChanges} from '../interface/simple_change';
 
 
 type Constructor<T> = new (...args: any[]) => T;
