+++ packages/core/test/BUILD.bazel
@@ -21,6 +21,7 @@ ts_library(
         "//packages/core",
         "//packages/core/src/di/interface",
         "//packages/core/src/interface",
+        "//packages/core/src/reflection",
         "//packages/core/src/util",
         "//packages/core/testing",
         "//packages/platform-browser",
